Fixed Address Issues
Restaurant page: Initialize adress list if it is empty so it does not take us to address page to pick an address

address page: we used to pick the last address in the list now we pick the first one which is the most latest one

Location Repo: Did not have a method to clear my local storage of previous addresses so i made one.

location controller: Added the cleanAddressHistory method to clear address from local memory on sign out. Also clear the placemarker so the name won't be wrong for couple seconds