Fixed notifications and address
Address had a problem. When the user first signed in the local storage was empty so I put the address in the local storage. Check location controller

build gradle was givving me errors so I did couple changes

Order_controller was giving me an error as well when the order fails so had to fix that as well

auth_repo had a logic error, that was not triggering the android notification, it was out of the if statement